Part 1 of the dap4 refactor. 1. Move all cdm dependent code into dap4/d4cdm 2. Modify various gradle files to support #1. 3. Misc. other mods to support #1.
