+++ dap4/d4cdm/src/main/java/dap4/cdm/CDMArrayAtomic.java
@@ -6,7 +6,6 @@ package   dap4.cdm;
 
 import dap4.core.dmr.*;
 import dap4.core.util.*;
-import dap4.cdmshared.CDMUtil;
 import dap4.dap4shared.*;
 import ucar.ma2.*;
