+++ dap4/d4cdm/src/main/java/dap4/cdm/CDMDataVariable.java
@@ -2,7 +2,7 @@
    See the LICENSE file for more information.
 */
 
-package dap4.servlet;
+package dap4.cdm;
 
 import dap4.core.data.DataVariable;
 import dap4.core.util.Slice;
