+++ dap4/d4tests/src/test/java/dap4/test/TestNc4Iosp.java
@@ -1,6 +1,6 @@
 package dap4.test;
 
-import dap4.servlet.CDMDSP;
+import dap4.cdm.CDMDSP;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
