+++ settings.gradle
@@ -6,8 +6,7 @@ include 'bufr'
 include 'cdm'
 include 'cdm-test'
 include 'clcommon'
-include 'dap4:d4cdmclient'
-include 'dap4:d4cdmshared'
+include 'dap4:d4cdm'
 include 'dap4:d4client'
 include 'dap4:d4core'
 include 'dap4:d4servletshared'
