+++ tds/build.gradle
@@ -19,6 +19,7 @@ dependencies {
     // DAP4 Dependencies (technically forward)
     compile project(":dap4:d4core")
     compile project(":dap4:d4servletshared")
+    compile project(":dap4:d4cdm")
 
     // Server stuff
     provided libraries["javax.servlet-api"]
